@@ -5,38 +5,36 @@ description = ""
 authors = ["Admin <admin@example.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.7"
-uvicorn = "^0.11.3"
-fastapi = "^0.54.1"
-python-multipart = "^0.0.5"
-email-validator = "^1.0.5"
-requests = "^2.23.0"
-celery = "^4.4.2"
-passlib = {extras = ["bcrypt"], version = "^1.7.2"}
-tenacity = "^6.1.0"
-pydantic = "^1.4"
-emails = "^0.5.15"
-raven = "^6.10.0"
-gunicorn = "^20.0.4"
-jinja2 = "^2.11.2"
-psycopg2-binary = "^2.8.5"
-alembic = "^1.4.2"
-pytest = "^5.4.1"
-python-jose = {extras = ["cryptography"], version = "^3.1.0"}
-asyncpg = "^0.22.0"
-httpx = "^0.17.1"
-SQLAlchemy = "^1.4.5"
+python = "^3.11"
+uvicorn = "^0.22.0"
+fastapi = "^0.100.0"
+python-multipart = "^0.0.6"
+email-validator = "^2.0.0"
+requests = "^2.31.0"
+celery = "^5.3.1"
+passlib = {extras = ["bcrypt"], version = "^1.7.4"}
+tenacity = "^8.2.2"
+pydantic = "^2.0.3"
+pydantic-settings = "^2.0.2"
+emails = "^0.6"
+gunicorn = "^20.1.0"
+jinja2 = "^3.1.2"
+alembic = "^1.11.1"
+python-jose = {extras = ["cryptography"], version = "^3.3.0"}
+asyncpg = "^0.27.0"
+httpx = "^0.24.1"
+SQLAlchemy = "^2.0.16"
 
 [tool.poetry.dev-dependencies]
-mypy = "^0.770"
-black = "^19.10b0"
-isort = "^4.3.21"
-autoflake = "^1.3.1"
-flake8 = "^3.7.9"
-pytest = "^5.4.1"
-sqlalchemy-stubs = "^0.3"
-pytest-cov = "^2.8.1"
-pytest-asyncio = "^0.14.0"
+mypy = "^1.3.0"
+black = "^23.3.0"
+isort = "^5.12.0"
+autoflake = "^2.1.1"
+flake8 = "^6.0.0"
+pytest = "^7.3.2"
+sqlalchemy-stubs = "^0.4"
+pytest-cov = "^4.1.0"
+pytest-asyncio = "^0.21.0"
 
 [tool.isort]
 multi_line_output = 3
