@@ -1,8 +1,10 @@
+import asyncio
 import logging
 
+from sqlalchemy import text
 from tenacity import after_log, before_log, retry, stop_after_attempt, wait_fixed
 
-from app.db.session import SessionLocal
+from app.db.session import get_session
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -17,11 +19,10 @@ wait_seconds = 1
     before=before_log(logger, logging.INFO),
     after=after_log(logger, logging.WARN),
 )
-def init() -> None:
+async def init() -> None:
     try:
-        # Try to create session to check if DB is awake
-        db = SessionLocal()
-        db.execute("SELECT 1")
+        async with get_session() as db:
+            await db.execute(text("SELECT 1"))
     except Exception as e:
         logger.error(e)
         raise e
@@ -29,7 +30,8 @@ def init() -> None:
 
 def main() -> None:
     logger.info("Initializing service")
-    init()
+    loop = asyncio.get_event_loop()
+    loop.run_until_complete(init())
     logger.info("Service finished initializing")
 
 
