@@ -6,10 +6,12 @@ from typing import Any, Dict, Optional
 import emails
 from emails.template import JinjaTemplate
 from jose import jwt
+from jose.exceptions import JWTError
 
 from app.core.config import settings
 
 
+# Should swap to async email library.
 def send_email(
     email_to: str,
     subject_template: str = "",
@@ -102,5 +104,5 @@ def verify_password_reset_token(token: str) -> Optional[str]:
     try:
         decoded_token = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])
         return decoded_token["email"]
-    except jwt.JWTError:
+    except JWTError:
         return None
